requirements.txt
# Core dependencies
openai>=1.0.0
PyYAML>=6.0.1

# Additional dependencies
PyMuPDF>=1.18.0